streamline import orders
