+++ pyat/at/physics/__init__.py
@@ -2,17 +2,18 @@
 Accelerator physics functions
 """
 # noinspection PyUnresolvedReferences
-from ..lattice import DConstant     # For backward compatibility
+# Avalailable here for backward compatibility
+from ..lattice import DConstant, Orbit, frequency_control
 from .amat import *
 from .energy_loss import *
+from .revolution import *
+from .harmonic_analysis import *
 from .orbit import *
 from .matrix import *
-from .revolution import *
 from .linear import *
 from .diffmatrix import find_mpole_raddiff_matrix
 from .radiation import *
 from .ring_parameters import *
-from .harmonic_analysis import *
 from .nonlinear import *
 from .fastring import *
 from .frequency_maps import fmap_parallel_track
